+++ ld_script.txt
@@ -263,6 +263,7 @@ SECTIONS {
         asm/code_807CABC.o(.text);
         src/targeting_flags.o(.text);
         asm/code_807CD9C.o(.text);
+        src/code_807CD9C.o(.text);
         src/weather.o(.text);
         asm/code_807E5AC.o(.text);
         src/code_807E5AC.o(.text);
