+++ src/options_menu.c
@@ -110,7 +110,8 @@ enum OptionsMenuStates {
     OPTIONS_MENU_PRE_HINT_SELECTION,
     OPTIONS_MENU_HINT_SELECTION,
     OPTIONS_MENU_DISPLAY_HINT,
-    OPTIONS_MENU_CONFIRM_NEW_OPTIONS = 7,
+    OPTIONS_MENU_UNKNOWN_6,
+    OPTIONS_MENU_CONFIRM_NEW_OPTIONS,
 };
 
 u32 sub_801DCC4(void)
